#127 - moved tests; fixed javadoc
