+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/features/FeaturesValidCrsParameterTransform.java
@@ -55,6 +55,8 @@ public class FeaturesValidCrsParameterTransform extends AbstractFeaturesCrs {
      *            id id of the collection, never <code>null</code>
      * @param collection
      *            the /collection object, never <code>null</code>
+     * @throws ParseException
+     *             if the geometry could not be parsed
      */
     @Test(description = "Implements A.2.1 Query, Parameter crs, Abstract Test 7 (Requirement /req/crs/crs-action), "
                         + "Geometries in the path /collections/{collectionId}/items", dataProvider = "collectionIdAndJson", dependsOnGroups = "crs-conformance")
@@ -95,6 +97,8 @@ public class FeaturesValidCrsParameterTransform extends AbstractFeaturesCrs {
      *            the /collection object, never <code>null</code>
      * @param crs
      *            the crs to test, never <code>null</code>
+     * @throws ParseException
+     *             if the geometry could not be parsed
      */
     @Test(description = "Implements A.2.1 Query, Parameter crs, Abstract Test 7 (Requirement /req/crs/crs-action), "
                         + "Transformed geometries in the path /collections/{collectionId}/items", dataProvider = "collectionIdAndJsonAndCrs", dependsOnGroups = "crs-conformance", dependsOnMethods = "verifyFeaturesPathGeometriesDefaultCrs")
